refactor: 줄일 수 있는 체이닝 제거 및 패키지 분리
