--- src/main/java/chess/domain/Rank.java
@@ -1,14 +1,0 @@
-package chess.domain;
-
-public enum Rank {
-
-    ONE,
-    TWO,
-    THREE,
-    FOUR,
-    FIVE,
-    SIX,
-    SEVEN,
-    EIGHT,
-    ;
-}
