+++ src/main/java/chess/domain/position/Rank.java
@@ -1,0 +1,14 @@
+package chess.domain.position;
+
+public enum Rank {
+
+    EIGHT,
+    SEVEN,
+    SIX,
+    FIVE,
+    FOUR,
+    THREE,
+    TWO,
+    ONE,
+    ;
+}
