+++ src/main/java/chess/dto/BoardOutput.java
@@ -1,6 +1,6 @@
 package chess.dto;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import chess.domain.piece.Piece;
 
 import java.util.Map;
